feat: validator 추가
피드백 반영하여 검증만을 담당하는 validator 추가
